+++ src/archive/ViewAccount/index.tsx
@@ -6,8 +6,6 @@ import React, {
 
 import styles from '../../../../styles/Account/Account.module.css';
 import { useAuth } from "../../lib/hooks/useAuth";
-
-import { initializeApp } from "firebase/app";
 import { getFirestore, 
     doc, 
     getDoc,
@@ -48,24 +46,9 @@ const accountData = [
     }}
 ]
 
-
-const firebaseConfig = {
-    apiKey: "AIzaSyCOnXDWQ369OM1lW0VC5FdYE19q1ug0_dc",
-    authDomain: "earmark-8d1d3.firebaseapp.com",
-    projectId: "earmark-8d1d3",
-    storageBucket: "earmark-8d1d3.appspot.com",
-    messagingSenderId: "46302537330",
-    appId: "1:46302537330:web:403eac7f28d2a4868944eb",
-    measurementId: "G-5474KY2MRV"
-};
-const app = initializeApp(firebaseConfig);
-const db = getFirestore(app);
-
-
 const ViewAccount = () => {
     const [docData, setDocData] = useState(null);
     const auth = useAuth();
-
 
     const getUserInfo = async () => {
         /* @ts-ignore */