+++ src/lib/hooks/useFirestore/index.tsx
@@ -1,33 +1,28 @@
 /* eslint-disable */
 import React, { useState, useEffect, useContext, createContext } from "react";
-import { initializeApp } from "firebase/app";
-import { getAuth, 
-  signInWithEmailAndPassword, 
-  createUserWithEmailAndPassword,
-  signOut,
-  sendPasswordResetEmail,
-  onAuthStateChanged,
-} from "firebase/auth";
 import { doc,
   setDoc,
   getFirestore,
 } from 'firebase/firestore'
 import { useAuth } from "../useAuth";
+import { initializeApp } from "firebase/app";
+import { getAnalytics, logEvent } from "firebase/analytics";
 
-// Paramater ordering: user_id, phone_number?, email?, first_name?, last_name?, account_id?, dob?, street?, city?, state?, zip?, username?, setup?
 import updateFirestoreUser from "../../firestore/updateFirestoreUser"
 
+
 const firebaseConfig = {
-  apiKey: "AIzaSyCOnXDWQ369OM1lW0VC5FdYE19q1ug0_dc",
-  authDomain: "earmark-8d1d3.firebaseapp.com",
-  projectId: "earmark-8d1d3",
-  storageBucket: "earmark-8d1d3.appspot.com",
-  messagingSenderId: "46302537330",
-  appId: "1:46302537330:web:403eac7f28d2a4868944eb",
-  measurementId: "G-5474KY2MRV"
+    apiKey: "AIzaSyCOnXDWQ369OM1lW0VC5FdYE19q1ug0_dc",
+    authDomain: "earmark-8d1d3.firebaseapp.com",
+    projectId: "earmark-8d1d3",
+    storageBucket: "earmark-8d1d3.appspot.com",
+    messagingSenderId: "46302537330",
+    appId: "1:46302537330:web:403eac7f28d2a4868944eb",
+    measurementId: "G-5474KY2MRV"
 };
-const app = initializeApp(firebaseConfig);
-const db = getFirestore(app);
+const firebaseApp = initializeApp(firebaseConfig);
+const db = getFirestore(firebaseApp);
+
 
 const firestoreContext = createContext({});
 // Provider component that wraps your app and makes auth object ...
@@ -36,15 +31,17 @@ export function ProvideFirestore({ children }) {
   const auth = useProvideFirestore();
   return <firestoreContext.Provider value={auth}>{children}</firestoreContext.Provider>;
 }
-// Hook for child components to get the auth object ...
-// ... and re-render when it changes.
-export const useFirestore = () => {
-  return useContext(firestoreContext);
-};
+
 
 function useProvideFirestore() {
   const [success, setSuccess] = useState(false);
   const auth = useAuth();
+  const app = firebaseApp;
+  let analytics;
+
+  if (app.name && typeof window !== 'undefined') {
+    analytics = getAnalytics(app);
+  }
 
   const test = () => {
     console.log('test func, inside the useFirestore hook')
@@ -140,7 +137,26 @@ function useProvideFirestore() {
     .catch((err) => {
       console.log('err updating category, ', err)
     })
-  }
+  };
+
+  const logSignIn = (method) => {
+    logEvent(analytics, 'signin', { 
+        method: method,
+    });
+  };
+
+  const logLinkInstitution = (institution) => {
+    logEvent(analytics, 'link_institution', { 
+        institution: institution,
+    });
+  };
+
+  const logTest = (test) => {
+    console.log('clicked test button', test)
+    logEvent(analytics, 'test', { 
+      test: test,
+    });
+  };
 
   useEffect(() => {
     setSuccess(false);
@@ -148,22 +164,31 @@ function useProvideFirestore() {
   }, [success])
 
   return {
+    app,
     success,
     test,
     createUserEntry,
     addBillingPlan,
     updateCategory,
     setupUserAccount,
+    logSignIn,
+    logLinkInstitution,
+    logTest,
   };
 }
 
-// TODO: setup interface so I don't need to add @ts-ignore above every call of this hook
-/*
-interface AuthContextType {
-  user: IUser;
-  signOut: () => void;
-  signIn: () => void;
+
+interface IProvideFirestore {
+  app: any;
+  success: boolean;
+  test: () => void;
+  createUserEntry: (user_id: string, phoneNumber: string, email: string, firstName: string, lastName: string) => void;
+  addBillingPlan: (user_id: string, plan: string) => void;
+  updateCategory: (category: string, transactionObjects: any) => void;
+  setupUserAccount: (user_id: string, dob: string, street: string, street2: string, city: string, state: string, zip: string, userId: string) => void;
+  logSignIn: (method: string) => void;
+  logLinkInstitution: (institution: string) => void;
+  logTest: (test: string) => void;
 }
 
-export const useAuth = () => React.useContext(AuthContext) as AuthContextType;
-*/
+export const useFirestore = () => useContext(firestoreContext) as IProvideFirestore;